Roadmap/voice-agent-pro/stt/requirements.txt
deepgram-sdk>=3.0.0
python-dotenv>=1.0.0
pyaudio>=0.2.11
websockets>=11.0.0
flask>=2.3.0
asyncio